implemented merge all functionality

+++ custom_modules/models/ProjectsModel.js
@@ -33,6 +33,12 @@ module.exports = class ProjectsModel extends AbstractModel {
 					if(err) {
 						controller.handleError(err);
 					}
+					var l = files.length;
+					while(l--) {
+						if(files[l].indexOf(".") == 0) {
+							files.splice(l, 1);
+						}
+					}
 					this.savedProjects = files.sort();
 					this._dispatchUpdate();
 				});
@@ -79,9 +85,11 @@ module.exports = class ProjectsModel extends AbstractModel {
 		this.fs.readdir(this.projectsDirectory, (err, files) => {
 			var l = files.length;
 			while(l--) {
-				var proj = this.fs.readJsonSync(this.projectsDirectory + '/' + files[l]);
-				if(proj && proj.project && proj.project.id == id) {
-					this.fs.removeSync(this.projectsDirectory + '/' + files[l]);
+				if(files[l].indexOf(".") != 0) {
+					var proj = this.fs.readJsonSync(this.projectsDirectory + '/' + files[l]);
+					if(proj && proj.project && proj.project.id == id) {
+						this.fs.removeSync(this.projectsDirectory + '/' + files[l]);
+					}
 				}
 			}
 		});
